Entrega final

+++ back_end/src/main/java/com/Xplored/Xplored/Model/Place/PlaceRepository.java
@@ -1,10 +1,9 @@
 package com.Xplored.Xplored.Model.Place;
 
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PlaceRepository extends JpaRepository<Place, Integer> {
+public interface PlaceRepository extends JpaRepository<Place, Long> {
 
 }
